Change documentation and add configfile argument
Configuration options are from previous commit documented directly in
example configuration file. So they don't have to be documented in
readme.

New argument configfile was added because if user uses default paths
for configuration files, it can be sometime unclear what file was used.
This prints path to used file.
Also as small feature, if no identifier is specified as argument neither
in configuration file help text is printed.